requirements.txt
# Spotify App Agent Template Dependencies

# Core dependencies
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
APScheduler>=3.10.4
SQLAlchemy>=2.0.23

# Spotify API
spotipy>=2.23.0

# Data analysis and visualization
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.0
numpy>=1.24.4

# Web dashboard
Flask>=3.0.0
Flask-SocketIO>=5.3.6
python-socketio>=5.10.0

# Image processing
Pillow>=10.1.0

# Development and testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
bandit>=1.7.0
safety>=2.3.0
pre-commit>=3.3.0

# Security and monitoring
cryptography>=41.0.8
prometheus-client>=0.19.0

# CLI enhancements
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=1.0.0

# Type stubs for better mypy support
types-requests>=2.28.0
types-PyYAML>=6.0.0
